Move back positions and velocities to f32, keeping reproductibility.

+++ exelixi/src/organism/position.rs
@@ -4,14 +4,14 @@ use crate::prelude::*;
 
 #[derive(Clone, Copy, Component, Debug)]
 pub struct Position {
-    pub x: i32,
-    pub y: i32,
+    pub x: f32,
+    pub y: f32,
     // Orientation in radian
     angle: f32,
 }
 
 impl Position {
-    pub fn new(x: i32, y: i32, angle: f32) -> Self {
+    pub fn new(x: f32, y: f32, angle: f32) -> Self {
         let mut pos = Self { x, y, angle: 0.0 };
         pos.set_angle(angle);
         pos
